+++ src/PipedEverything/Properties/AssemblyInfo.cs
@@ -2,5 +2,5 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("PipedEverything")]
 [assembly: AssemblyProduct("PipedEverything")]
-[assembly: AssemblyVersion("1.0.27.13")]
-[assembly: AssemblyFileVersion("1.0.27.13")]
+[assembly: AssemblyVersion("1.0.27.1")]
+[assembly: AssemblyFileVersion("1.0.27.1")]