+++ src/PipedEverything/_FumiKMod.cs
@@ -13,7 +13,21 @@ namespace PipedEverything
     public class FumiKMod : KMod.UserMod2
     {
         public const string ModName = "PipedEverything";
-        public static Harmony instance;
+        public static Harmony? instance;
+
+        private static float _solidMaxMass = float.NaN;
+        public static float SolidMaxMass
+        {
+            get
+            {
+                if (_solidMaxMass == float.NaN)
+                {
+                    var warpStorage = Assets.TryGetPrefab(WarpConduitSenderConfig.ID).GetComponent<WarpConduitSender>().solidStorage;
+                    _solidMaxMass = warpStorage.capacityKg / 5f;
+                }
+                return _solidMaxMass;
+            }
+        }
 
         public override void OnLoad(Harmony harmony)
         {